Small improvements (#78)
* Increase num vehicles in bin packing heuristic

* Fix fitness calculation

* Change nbIter default 20k to 10k

+++ hgs_vrptw/include/Population.h
@@ -18,7 +18,7 @@ class Population
     struct IndividualWrapper
     {
         std::unique_ptr<Individual> indiv;
-        size_t fitness;
+        double fitness;
 
         bool operator<(IndividualWrapper const &other) const
         {
